+++ Cadmus.Philology.Parts/MspOperation.cs
@@ -1,4 +1,5 @@
-﻿using Fusi.Tools.Text;
+﻿/*
+using Fusi.Tools.Text;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Text.RegularExpressions;
@@ -433,3 +434,4 @@ public enum MspOperator
     /// <summary>Swap</summary>
     Swap
 }
+*/